pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lore-kb"
version = "0.1.0"
description = "Personal ML research knowledge base with evolving agent"
requires-python = ">=3.10"
dependencies = [
    "peft>=0.10.0",
    "trl>=0.8.0",
    "sentence-transformers>=3.0.0",
    "bitsandbytes>=0.42.0",
    "rank-bm25>=0.2.2",
    "pdfminer.six>=20221105",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "sqlitedict>=2.1.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "networkx>=3.3",
    "httpx>=0.27.0",
]

[project.scripts]
wiki = "lore.cli.wiki_cli:main"
lore = "lore.cli.wiki_cli:main"
lore-ingest = "lore.cli.wiki_cli:ingest_main"
lore-search = "lore.cli.wiki_cli:search_main"
lore-absorb = "lore.cli.wiki_cli:absorb_main"
lore-rebuild-index = "lore.cli.wiki_cli:rebuild_index_main"
lore-health = "lore.cli.wiki_cli:health_main"
lore-status = "lore.cli.wiki_cli:status_main"
lore-render = "lore.cli.wiki_cli:render_main"
lore-query = "lore.cli.wiki_cli:query_main"
lore-cleanup = "lore.cli.wiki_cli:cleanup_main"
lore-reorganize = "lore.cli.wiki_cli:reorganize_main"
lore-train = "lore.cli.evolve_cli:main"
lore-web = "lore.web.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lore"]

[tool.hatch.build]
sources = ["src"]
